refactor(calculator): 완성 코드 메인 패키지 이동
