+++ src/main/java/calculator/strategy/DynaminTextDelimiter.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.strategy;
 
 import java.util.Arrays;
 import java.util.List;
