+++ src/main/java/calculator/vo/PositiveInt.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.vo;
 
 public record PositiveInt(int value) {
